+++ google-app-engine/base_handler.py
@@ -5,7 +5,7 @@
     All handlers must be subclasses of `Handler` from the fwerks `fwerks.py`
     module. See `fwerks.py` for documentation regarding request handlers.
 
-    @author Kris Walker <eixxauth@gmail.com>
+    @author Kris Walker <kixxauth@gmail.com>
     @copyright (c) 2010 by The Fireworks Project.
     @license MIT, see MIT-LICENSE for more details.
 """
@@ -18,10 +18,14 @@ from werkzeug import BaseResponse, CommonResponseDescriptorsMixin, ETagResponseM
 from werkzeug import UserAgent, cached_property
 
 from openid.consumer.consumer import Consumer, FAILURE, SUCCESS, CANCEL
+from openid.yadis.discover import DiscoveryFailure
 import aeoid_store
 
+import utils
 import fwerks
 import dstore
+
+NO_CACHE_HEADER = utils.NO_CACHE_HEADER
 
 Handler = fwerks.Handler
 User =    fwerks.User
@@ -41,9 +45,6 @@ beaker_session_configs = {
 
 # ### The URL query param used for the federated identifier.
 AUTH_REQUEST_FIELD = 'federated_id'
-
-# ### The URL query param signalling a federated login process is active.
-AUTH_PROCESS_FIELD = 'federated_authentication'
 
 
 class Request(BaseRequest, CommonRequestDescriptorsMixin, AcceptMixin, ETagRequestMixin):
@@ -103,21 +104,21 @@ class BaseHandler(Handler):
             # expired session, creating a new one or prompting another login.
             return User(session)
 
-        if self.values.get(AUTH_PROCESS_FIELD) is None:
-            return
-
         # This is a redirected federated login attempt.  If this raises an
         # exception it will be caught by the fwerks dispatcher.
-        logging.warn('self.request.url: %s', self.request.url)
+        # TODO: Find a nicer way to handle exceptions.
         auth_reponse = Consumer( session
                                , aeiod_store.AppEngineStore()
                                ).complete(self.request.args, self.request.url)
 
         if response.status == SUCCESS:
             session['authenticated'] = True
+            session.save()
             return User(session)
 
-        elif response.status in (FAILURE, CANCEL):
+        session.delete()
+
+        if response.status in (FAILURE, CANCEL):
             return
         else:
             logging.error('Unexpected error in OpenID authentication: %s', response)
@@ -260,7 +261,7 @@ class AuthRequestHandler(BaseHandler):
     def get(self):
         federated_id = self.request.args.get(AUTH_REQUEST_FIELD)
         if not federated_id:
-            response = Response()
+            response = Response('missing %s field\n'% AUTH_REQUEST_FIELD)
             # Since we only care about Ajax requests we don't have to do
             # anything more informative than a 409.
             response.status_code = 409
@@ -268,22 +269,25 @@ class AuthRequestHandler(BaseHandler):
 
         session = self.environ.get(BEAKER_ENV_KEY)
 
-        # If this raises an exception it will be caught by the fwerks dispatcher.
-        auth_request = Consumer( session
-                               , aeiod_store.AppEngineStore()
-                               ).begin(federated_id)
+        try:
+            auth_request = Consumer( session
+                                   , aeoid_store.AppEngineStore()
+                                   ).begin(federated_id)
+        except DiscoveryFailure:
+            response = Response('discovery failure\n')
+            # Since we only care about Ajax requests we don't have to do
+            # anything more informative than a 409.
+            response.status_code = 409
+            return self.respond(response)
 
-        cont = list(urlparse.urlparse(self.args.get('continuation', '/')))
-        qstr = AUTH_PROCESS_FIELD +'=true'
-        cont[4] = cont[4] and (cont[4] + ('&'+ qstr)) or ('?'+ qstr)
+        # If this raises a different exception it will be caught by the fwerks dispatcher.
+        # TODO: We need a nicer way of handling exceptions here.
 
-        return_to = urlparse.urljoin( self.request.host
-                                    , urlparse.urlunparse(cont)
-                                    )
+        host_url = self.request.host_url
+        cont = self.request.args.get('continuation', '/')
+        return_to = urlparse.urljoin(host_url, cont)
 
-        logging.warn('return_to: %s', return_to)
-
-        redirect_url = auth_request.redirectURL(self.request.host, return_to)
+        redirect_url = auth_request.redirectURL(host_url, return_to)
         # The Ajax request wants the redirect url as plain text, not an actual
         # HTTP redirect.
         response = self.respond(Response(redirect_url))